Imported typings in toaster file. Fixed case typo
