--- EchoFramework/src/types/toaster.ts
@@ -1,8 +1,0 @@
-import { ToasterMessage } from "./toasterMessage";
-
-export interface Toaster {
-    toastNumber: number;
-    toast: ToasterMessage;
-    numberOfToasts: number;
-    onClose?: () => void;
-}
